// x-pack/solutions/observability/plugins/apm/common/waterfall/typings.d.ts
import type { Exception } from '../../typings/es_schemas/raw/error_raw';
import type { EventOutcome } from '../../typings/es_schemas/raw/fields/event_outcome';
import type { SpanLink } from '../../typings/es_schemas/raw/fields/span_links';
import type { TimestampUs } from '../../typings/es_schemas/raw/fields/timestamp_us';
import type { AgentName } from '../../typings/es_schemas/ui/fields/agent';
export interface WaterfallTransaction {
    timestamp: TimestampUs;
    trace: {
        id: string;
    };
    service: {
        name: string;
        environment?: string;
    };
    agent: {
        name: AgentName;
    };
    event?: {
        outcome?: EventOutcome;
    };
    parent?: {
        id?: string;
    };
    processor: {
        event: 'transaction';
    };
    transaction: {
        duration: {
            us: number;
        };
        id: string;
        name: string;
        type: string;
        result?: string;
    };
    faas?: {
        coldstart?: boolean;
    };
    span?: {
        id?: string;
        links?: SpanLink[];
    };
}
export interface WaterfallSpan {
    timestamp: TimestampUs;
    trace: {
        id: string;
    };
    service: {
        name: string;
        environment?: string;
    };
    agent: {
        name: AgentName;
    };
    event?: {
        outcome?: EventOutcome;
    };
    parent?: {
        id?: string;
    };
    processor: {
        event: 'span';
    };
    span: {
        id: string;
        type: string;
        subtype?: string;
        action?: string;
        name: string;
        composite?: {
            count: number;
            sum: {
                us: number;
            };
            compression_strategy: string;
        };
        sync?: boolean;
        duration: {
            us: number;
        };
        links?: SpanLink[];
        destination?: {
            service?: {
                resource?: string;
            };
        };
    };
    transaction?: {
        id?: string;
    };
    child?: {
        id: string[];
    };
}
/**
 * @deprecated Use Error instead
 */
export interface WaterfallError {
    id: string;
    timestamp: TimestampUs;
    trace?: {
        id?: string;
    };
    transaction?: {
        id?: string;
    };
    parent?: {
        id?: string;
    };
    span?: {
        id?: string;
    };
    eventName?: string;
    error: {
        id: string;
        log?: {
            message: string;
        };
        exception?: Exception[];
        grouping_key: string;
    };
    service: {
        name: string;
    };
}
export type IWaterfallGetRelatedErrorsHref = (docId: string) => string;
export type WaterfallGetServiceBadgeHref = (serviceName: string) => string;